refactor: move gateway tests to module file
